move restricted method test into own file
