+++ test/unit/app/controllers/permissions/permissions-controller-test.js
@@ -8,9 +8,6 @@ import {
   WALLET_PREFIX,
 } from '../../../../../app/scripts/controllers/permissions/enums'
 
-import getRestrictedMethods
-  from '../../../../../app/scripts/controllers/permissions/restrictedMethods'
-
 import {
   PermissionsController,
   addInternalMethodPrefix,
@@ -1189,30 +1186,6 @@ describe('permissions controller', function () {
       const str = 'foo'
       const res = addInternalMethodPrefix(str)
       assert.equal(res, WALLET_PREFIX + str, 'should prefix correctly')
-    })
-
-    it('eth_accounts: restricted method failure', async function () {
-      const restrictedMethods = getRestrictedMethods({
-        getKeyringAccounts: async () => {
-          throw new Error('foo')
-        },
-      })
-
-      const res = {}
-      restrictedMethods.eth_accounts.method(null, res, null, (err) => {
-
-        const fooError = new Error('foo')
-
-        assert.deepEqual(
-          err, fooError,
-          'error should be as expected'
-        )
-
-        assert.deepEqual(
-          res, { error: fooError },
-          'response should have correct error and no message'
-        )
-      })
     })
   })
 })
